LPS-153063 Rename StyleBookEditorContext and move the context to the corresponding folder
