+++ modules/apps/style-book/style-book-web/src/main/resources/META-INF/resources/js/style-book-editor/Undo.js
@@ -22,7 +22,7 @@ import {
 	useOnUndo,
 	useRedoHistory,
 	useUndoHistory,
-} from './StyleBookContext';
+} from './contexts/StyleBookEditorContext';
 
 export default function Undo() {
 	const onUndo = useOnUndo();
